# legacy/docs/swagger.yaml
basePath: /
definitions:
  bureau-backend_internal_core_domain.HealthStatus:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  bureau-backend_internal_core_domain.User:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        description: For input only
        type: string
      updated_at:
        type: string
    type: object
  internal_adapters_handler_http.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_adapters_handler_http.LoginResponse:
    properties:
      token:
        type: string
    type: object
  internal_adapters_handler_http.RegisterRequest:
    properties:
      active:
        type: boolean
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: This is the backend API for Bureau
  title: Bureau Backend API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password to get JWT token
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_handler_http.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapters_handler_http.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/bureau-backend_internal_core_domain.HealthStatus'
      summary: Login user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with name, email and password
      parameters:
      - description: User Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_handler_http.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/bureau-backend_internal_core_domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/bureau-backend_internal_core_domain.HealthStatus'
      summary: Register a new user
      tags:
      - auth
swagger: "2.0"